+++ plugin-api/beta/client/hocs/index.js
@@ -27,7 +27,5 @@ export {
   withSetCommentStatus,
   withChangePassword,
   withChangeUsername,
-  withUpdateEmailAddress,
-  withAttachLocalAuth,
 } from 'coral-framework/graphql/mutations';
 export { compose } from 'recompose';
